+++ src/Entity/Versement.php
@@ -30,7 +30,7 @@ class Versement
     #[ORM\Column(type:"integer", nullable:true) ]
     private $numero;
 
-    #[ORM\Column(type:"integer") ]
+    #[ORM\Column(type:"float") ]
     private $montant;
 
     #[ORM\ManyToOne(targetEntity:Banque::class, inversedBy:"versements") ]
@@ -72,12 +72,12 @@ class Versement
         return $this;
     }
 
-    public function getMontant(): ?int
+    public function getMontant(): ?float
     {
         return $this->montant;
     }
 
-    public function setMontant(int $montant): self
+    public function setMontant(float $montant): self
     {
         $this->montant = $montant;
